requirements.txt
# ABOUTME: Python dependencies for advanced typing pattern analyzer
# Core dependencies
pynput>=1.7.6
pyobjc-framework-Cocoa>=9.0
pyobjc-framework-Quartz>=9.0
pyobjc-framework-ApplicationServices>=9.0

# Data analysis and visualization
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
scipy>=1.10.0

# Configuration and serialization
PyYAML>=6.0
pydantic>=2.0.0

# Development and testing
pytest>=7.3.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.3.0

# Performance and utilities
psutil>=5.9.0
click>=8.1.0
rich>=13.0.0
tqdm>=4.65.0

# Optional ML dependencies
scikit-learn>=1.2.0